requirements.txt
slack-sdk
pygsheets~=2.0.0
python-dateutil~=2.8.0
PyYAML==6.0
requests
Quart==0.15.1